additon sub,multipliaction and division through function has been added

+++ addition ,substraction through function/main.cpp
@@ -33,7 +33,7 @@ int mul(int a ,int b)
 {
     return (a*b);
 }
-int div(int a ,int b)
+float div(int a ,int b)
 {
     return (a/b);
 }
